Gui v4.0.0 to develop

See merge request sparroweiss/trivia-project!21

+++ Trivia Project/Trivia Project/IDataBase.h
@@ -32,7 +32,7 @@ struct
 	int _totalAnswers;
 	int _correctAnswers;
 	int _numOfGames;
-	int _answersTime;
+	float _answersTime;
 }typedef Statistic;
 
 struct
